requirements.txt
PyPDF2==3.0.0
pdfplumber==0.8.0
pillow==9.0.0
pandas==2.0.3
